refactor: make the input handling use a more explicit state machine
**Issue:**

Currently, the implementation for input handling contains a few states
the plugin can be in. It's difficult to realize this because of the way
the code is structured.

It's not clear what variables are in which state, and it's difficult to
see what transitions between states are possible.

**Solution:**

Refactor the input handling code to use a more explicit state machine to
solve these issues. Now each state is a function, and sharing
unnecessary state is reduced.