docs: yazi.nvim instructions explain how to avoid update spam
**Issue:**

As part of ensuring the plugin works with the latest yazi changes, lots
of automatic package updates are done almost daily. This clutters the
update history and makes it hard to see important changes.

**Solution:**

I enabled automatic releases whenever user-facing changes are merged
(such as new features and bug fixes) in
https://github.com/mikavilpas/easyjump.yazi/pull/313

To avoid frequent package updates, yazi.nvim users can now set
`version = "*"` in their configuration. This way their
[lazy.nvim](https://lazy.folke.io/) will only suggest updating if
user-facing changes are made.

This solves the issue, but as far as I know it cannot be done for
`ya pkg` at this time.